+++ python/tvm/s_tir/data_layout.py
@@ -23,7 +23,7 @@ from tvm.runtime import Object
 from . import _ffi_api
 
 
-@tvm_ffi.register_object("tir.Layout")
+@tvm_ffi.register_object("s_tir.Layout")
 class Layout(Object):
     """Layout is composed of upper cases, lower cases and numbers,
     where upper case indicates a primal axis and
@@ -81,7 +81,7 @@ class Layout(Object):
         return _ffi_api.LayoutFactorOf(self, axis)  # type: ignore
 
 
-@tvm_ffi.register_object("tir.BijectiveLayout")
+@tvm_ffi.register_object("s_tir.BijectiveLayout")
 class BijectiveLayout(Object):
     """Bijective mapping for two layouts (src-layout and dst-layout).
     It provides shape and index conversion between each other.
