+++ python/tvm/tir/__init__.py
@@ -20,7 +20,6 @@ from tvm.ir import PrimExpr
 from tvm.runtime import const
 
 from .buffer import Buffer, decl_buffer, DataProducer
-from .data_layout import Layout, BijectiveLayout, bijective_layout, layout
 from .expr import convert
 from .expr import Var, SizeVar, Reduce, FloatImm, IntImm, StringImm, Cast
 from .expr import Add, Sub, Mul, Div, Mod, FloorDiv, FloorMod
