+++ src/contrib/msc/core/ir/graph_builder.h
@@ -29,7 +29,7 @@
 #include <tvm/relax/expr.h>
 #include <tvm/relax/expr_functor.h>
 #include <tvm/runtime/tensor.h>
-#include <tvm/tir/data_layout.h>
+#include <tvm/s_tir/data_layout.h>
 
 #include <set>
 #include <stack>
