+++ src/relax/transform/infer_amp_utils.h
@@ -29,7 +29,6 @@
 #include <tvm/relax/expr.h>
 #include <tvm/relax/nested_msg.h>
 #include <tvm/relax/op_attr_types.h>
-#include <tvm/tir/data_layout.h>
 
 #include <unordered_map>
 #include <unordered_set>
