+++ src/s_tir/analysis/sblock_buffer_access_lca_detector.cc
@@ -18,7 +18,7 @@
  */
 
 /*!
- * \file tir/analysis/buffer_access_lca_detector.cc
+ * \file s_tir/analysis/sblock_buffer_access_lca_detector.cc
  * \brief Detect the lowest common ancestor(LCA) of buffer access
  */
 
@@ -349,7 +349,7 @@ ffi::Map<Buffer, ffi::Optional<Stmt>> DetectBufferAccessLCA(const PrimFunc& func
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("tir.analysis.detect_buffer_access_lca", DetectBufferAccessLCA);
+  refl::GlobalDef().def("s_tir.analysis.detect_buffer_access_lca", DetectBufferAccessLCA);
 }
 }  // namespace tir
 }  // namespace tvm
