+++ src/tir/analysis/stmt_finding.cc
@@ -64,92 +64,5 @@ const PrimFuncNode* FindEntryFunc(const IRModule& mod, GlobalVar* result_g_var) 
   return nullptr;
 }
 
-Stmt GetEnclosingLoop(const SBlockNode* block, Stmt func_body) {
-  struct GetRootSeqStmt : public StmtVisitor {
-    void VisitStmt_(const SeqStmtNode* seq) override { result = seq; }
-    const SeqStmtNode* result;
-  };
-
-  struct BlockFinder : public StmtVisitor {
-    explicit BlockFinder(const SBlockNode* tgt) : target(tgt) {}
-
-    void VisitStmt_(const SBlockNode* block) override {
-      if (block == target) {
-        found = true;
-      }
-    }
-
-    const SBlockNode* target;
-    bool found = false;
-  };
-
-  GetRootSeqStmt seq_finder;
-  seq_finder(func_body);
-
-  ICHECK(seq_finder.result);
-
-  for (auto stmt : seq_finder.result->seq) {
-    if (stmt->IsInstance<ForNode>()) {
-      BlockFinder finder(block);
-      finder(stmt);
-      if (finder.found) {
-        return stmt;
-      }
-    }
-  }
-
-  LOG(FATAL) << "Enclosing loop not found for a block " << ffi::GetRef<SBlock>(block);
-  TVM_FFI_UNREACHABLE();
-}
-
-const SBlockNode* FindAnchorBlock(const IRModule& mod) {
-  struct ReductionSBlockCollector : public StmtVisitor {
-    void VisitStmt_(const SBlockNode* block) override {
-      if (block->init) {
-        blocks.push_back(block);
-      }
-      StmtVisitor::VisitStmt(block->body);
-    }
-    std::vector<const SBlockNode*> blocks;
-  };
-
-  if (auto prim_func = FindEntryFunc(mod, nullptr)) {
-    ReductionSBlockCollector collector;
-    collector(prim_func->body);
-
-    const auto& candidates = collector.blocks;
-
-    if (candidates.empty()) {
-      return nullptr;
-    } else if (candidates.size() == 1) {
-      return candidates[0];
-    }
-
-    double best_flops = -1;
-    int best_idx = 0;
-    for (size_t i = 0; i < candidates.size(); ++i) {
-      auto loop = GetEnclosingLoop(candidates[i], prim_func->body);
-      auto flops = EstimateTIRFlops(loop);
-      if (flops > best_flops) {
-        best_flops = flops;
-        best_idx = i;
-      }
-    }
-    return candidates[best_idx];
-  }
-  return nullptr;
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("tir.analysis.find_anchor_sblock", [](const IRModule& mod) {
-    auto ret = FindAnchorBlock(mod);
-    if (ret) {
-      return ffi::Optional<SBlock>(ffi::GetRef<SBlock>(ret));
-    }
-    return ffi::Optional<SBlock>(std::nullopt);
-  });
-}
-
 }  // namespace tir
 }  // namespace tvm
