+++ src/tir/ir/data_type_rewriter.cc
@@ -22,9 +22,13 @@
  * \brief Rewrite the data type of expressions.
  */
 
+#include "data_type_rewriter.h"
+
 #include <tvm/tir/builtin.h>
-#include <tvm/tir/data_type_rewriter.h>
 #include <tvm/tir/op.h>
+
+#include <algorithm>
+#include <utility>
 
 #include "./functor_common.h"
 #include "tvm/ir/expr.h"
