+++ tests/python/s_tir/base/test_tir_te_extern_primfunc.py
